backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy==2.0.48
pydantic[email]==2.12.5
python-jose==3.5.0
passlib==1.7.4
python-dotenv==1.2.2